requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
psycopg2-binary>=2.9.9,<3.0
cloudinary>=1.41.0,<2.0
dj-database-url>=2.1.0,<3.0
python-dotenv>=1.0.1,<2.0
google-auth>=2.27.0,<3.0
requests>=2.31.0,<3.0
django-cloudinary-storage>=0.3.0
django-cors-headers>=4.3.0
